Add ability to configure multiple kibana dashboards
Introduce the concept of namespace, which allows to run multiple kibana
dashboards against the same elasticsearch instance, in which dashboard
and indexes have been configured within the given namespace to avoid
collision.

Please note that this change does not prevent kibana dashboards to have
visibility on elasticsearch indexes beyond the namespace they currently
are configured for.

Authorization will be added in future releases.

Feature: Issue 9865
Change-Id: I02938a0c4ed38d84a47d5cc181443807e66c5e4a