Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > Do we *really* need to have different kibana instances? only for having different dashboards? Can't we just have different dashboards pointing to different indexes?
> 
> Let me try to clarify this.
> Issue 9865 states:
> 
> -------
> Allow to publish and compose Kibana dashboards for different sets of projects by leveraging the ability to store the dashboard definitions in different index.
> 
> Example:
> - Kibana index: kibana-gerrit
> - ES index: gerrit
> 
> That will have all the dashboards associated to Gerrit
> 
> - Kibana index: kibana-spdk
> - ES index: spdk
> ---------
> 
> The fact is that you cannot have more than one kibana index per kibana instance.
> So you cannot have the same kibana instance configured with "kibana-gerrit" and "kibana-spdk".
> 
> In other words, you cannot "store the dashboard definitions in different index.", all dashboard definitions will be stored in the same kibana index.

OK, if kibana can have only one index, what's the value of having a different index name if there is no conflict at all? Then you would just use another Kibana instance to serve the unique dashboard.

> What you _can_ do is to have different dashboards (mind, stored in the same kibana index), having different index-patterns, that is, reading from different elasticsearch indexes.

Yes, that would be a lot better. If Kibana could have different indexes, then yes it would have been a solution. Otherwise having different index names doesn't really solve the issues of having different dashboards on the same Kibana.

> This approach provides a solution whereby you get multiple kibana indexes, because you have different kibana instances. Some might prefer this because they can have different kibana settings for different audiences, different plugins depending on data, different security levels, etc.

Not following how that relates to the fact of having the analytics on Gerrit data. Kibana is only the "implementation details" of the Gerrit analytics data, I would not suggest people to start building on top of it plenty of other plugins, configurations etc.

Ideally, Kibana could be completely masked through a reverse proxy in Gerrit that shows *only* the graph and nothing else.

> I agree we could/should also provide a solution where we can configure multiple dashboards, within the same kibana index (for different index-pattern).

Yes, that was the purpose.

> I am happy, to address this in another changeset

If this change is useful for the target solution, we can keep it. Otherwise, it's just other code and entropy that we will have removed anyway in the future :-(

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>